+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add support of 16-bit BMP.
 - Add `IMGPROXY_NEW_RELIC_LABELS` config.
 - Add support of JPEG files with differential Huffman coding or arithmetic coding.
+- Add `IMGPROXY_PREFERRED_FORMATS` config.
 
 ### Fix
 - Fix trimming of CMYK images.